// src/app/api/models/FormationResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormationResponseDTO = {
    id?: string;
    company?: string | null;
    title?: string | null;
    city?: string | null;
    country?: string | null;
    startAt?: string;
    endAt?: string;
    userId?: string | null;
};