// src/app/api/models/FormationCreateDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormationCreateDTO = {
    company: string;
    title: string;
    city?: string | null;
    country?: string | null;
    startAt: string;
    endAt: string;
};