// src/app/api/models/FormationUpdateDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormationUpdateDTO = {
    id: string;
    company: string;
    title: string;
    city?: string | null;
    country?: string | null;
    startAt: string;
    endAt: string;
};